custom_serving_image/requirements.txt
torch
torchvision==0.15.2
Flask==2.0.1
gunicorn==21.2.0
gcsfs==2023.6.0
python-dotenv==1.0.0
numpy
tqdm==4.61.1
einops==0.6.1
timm==0.4.12
scipy
yacs==0.1.8
nibabel==3.2.1
